requirement.txt
certifi==2020.12.5
MiniSom==2.2.8
numpy==1.18.1
Pillow==8.1.2
PySimpleGUI==4.38.0
